requirements.txt
streamlit 
pandas 
scikit-learn
seaborn 
matplotlibe
